Add files via upload
Applications of Blockchain Beyond Cryptocurrencies - Value, Systems, and Challenges

Blockchain technology is most widely known for enabling the creation of crypto-currencies, with some of the most notable examples being Bitcoin and Ethereum. However, since their creation, crypto-currencies have primarily operated as a medium for the storage of value and have seen only peripheral use as a traditional currency. That said, the underlying technology of blockchain could and should have utility in many other domains. In our paper, we have surveyed the spectrum of blockchain solutions implemented in various industries. We’ve provided a particular focus on applications beyond currencies, looking at the finance industry, healthcare, and politics to understand what systems are used and how the technology will change the current techniques used in the industry.